fix not showing log in panel when user that were not logged in click the booking button
