+++ app.py
@@ -23,4 +23,4 @@ def booking():
 def thankyou():
 	return render_template("thankyou.html")
 
-app.run(host = "0.0.0.0", port = 3000, debug=True)
+app.run(host = "0.0.0.0", port = 3000)
