+++ templates/attraction.html
@@ -105,7 +105,7 @@
         <footer id="scroll-bottom-detector">
             <p class="body-bold" >COPYRIGHT © 2021 台北一日遊</p>
         </footer>
-        <script src="{{url_for('static', filename = 'scripts/attraction.js')}}"></script>
-        <script src="{{url_for('static', filename = 'scripts/userAuthentication.js')}}"></script>
+        <script type="module" src="{{url_for('static', filename = 'scripts/attraction.js')}}"></script>
+        <script type="module" src="{{url_for('static', filename = 'scripts/userAuthentication.js')}}"></script>
     </body>
 </html>
