+++ templates/booking.html
@@ -93,6 +93,6 @@
             <p class="body-bold" >COPYRIGHT © 2021 台北一日遊</p>
         </footer>
         <script src="{{url_for('static', filename = 'scripts/booking.js')}}"></script>
-        <script src="{{url_for('static', filename = 'scripts/userAuthentication.js')}}"></script>
+        <script type="module" src="{{url_for('static', filename = 'scripts/userAuthentication.js')}}"></script>
     </body>
 </html>
